pyproject.toml
[tool.poetry]
name = "herrschaft-der-asche"
version = "0.1.0"
description = "A simple text adventure"
authors = ["Der Mann mit Hut <dermannmithut@metstuebchen.de>"]
packages = [
  { include = "engine" },
  { include = "game" }
]

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
pyyaml = "^6.0"
pydantic = "^2.0"
requests = "^2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-cov = "^5.0"
pyright = "^1.1"
ruff = ">=0.12"

[tool.poetry.scripts]
"herrschaft-der-asche" = "game.main:run_cli"

[tool.ruff]
line-length = 140
target-version = "py312"
src = ["engine", "game", "tests"]

[tool.ruff.lint]
select = [
  "E", "F", "B", "I", "C4", "N", "Q", "S", "YTT", "DTZ", "UP", "SIM", "ARG"
]
ignore = ["S101", "S105"]

[tool.pyright]
include = ["engine", "game", "tests"]
pythonVersion = "3.12"
#typeCheckingMode = "strict"
#reportMissingImports = true
#reportUnusedImport = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
